game mode UI change
I implement the game mode choice UI, now they all integration on the same page with instruction. And improved the appearance of the scoring, using color to distinguish the importance. And use red to strong the present of the rule different. In the end, create a forfeit button to help forfeit.